GP-5887 moved some threading support into Generic
